Move reset functionality to drawer
